config: Cast price_stop_level to (float)

+++ config/EURUSD_M1.h
@@ -5,9 +5,7 @@
 
 // Defines indicator's parameter values for the given pair symbol and timeframe.
 struct ACParams_M1 : ACParams {
-  ACParams_M1() : ACParams(indi_ac_defaults, PERIOD_M1) {
-    shift = 0;
-  }
+  ACParams_M1() : ACParams(indi_ac_defaults, PERIOD_M1) { shift = 0; }
 } indi_ac_m1;
 
 // Defines strategy's parameter values for the given pair symbol and timeframe.
@@ -22,7 +20,7 @@ struct Stg_AC_Params_M1 : StgParams {
     signal_close_method = 0;
     signal_close_level = (float)0.1;
     price_stop_method = 0;
-    price_stop_level = 0.1;
+    price_stop_level = (float)0.1;
     tick_filter_method = 1;
     max_spread = 0;
   }
