+++ config/EURUSD_M15.h
@@ -5,9 +5,7 @@
 
 // Defines indicator's parameter values for the given pair symbol and timeframe.
 struct ACParams_M15 : ACParams {
-  ACParams_M15() : ACParams(indi_ac_defaults, PERIOD_M15) {
-    shift = 0;
-  }
+  ACParams_M15() : ACParams(indi_ac_defaults, PERIOD_M15) { shift = 0; }
 } indi_ac_m15;
 
 // Defines strategy's parameter values for the given pair symbol and timeframe.
@@ -22,7 +20,7 @@ struct Stg_AC_Params_M15 : StgParams {
     signal_close_method = 0;
     signal_close_level = (float)0.1;
     price_stop_method = 0;
-    price_stop_level = 0.1;
+    price_stop_level = (float)0.1;
     tick_filter_method = 1;
     max_spread = 0;
   }
